Refactor model tutorials into a package structure
